+++ simbad/command_line/simbad_morda.py
@@ -70,10 +70,9 @@ def main():
         logger.info("No results found - full search was unsuccessful")
 
     if args.output_pdb and args.output_mtz:
-        run_dir = os.path.join(args.work_dir, 'morda')
-        csv = os.path.join(run_dir, 'morda_mr.csv')
+        csv = os.path.join(args.work_dir, 'morda', 'morda_mr.csv')
         result = simbad.util.result_by_score_from_csv(csv, 'final_r_free', ascending=True)
-        simbad.util.output_files(run_dir, result, args.output_pdb, args.output_mtz)
+        simbad.util.output_files(args.work_dir, result, args.output_pdb, args.output_mtz)
 
     stopwatch.stop()
     logger.info("All processing completed in %d days, %d hours, %d minutes, and %d seconds",