+++ simbad/util/__init__.py
@@ -4,7 +4,6 @@ __author__ = "Adam Simpkin, Felix Simkovic & Jens Thomas"
 __date__ = "05 May 2017"
 __version__ = "1.0"
 
-import glob
 import logging
 import os
 import pandas as pd
@@ -22,9 +21,9 @@ logger = logging.getLogger(__name__)
 def output_files(run_dir, result, output_pdb, output_mtz):
     """Return output pdb/mtz from best result in result obj"""
     pdb_code = result[0]
-    stem = os.path.join(run_dir, 'mr_search', pdb_code, 'mr', '*', 'refine')
-    input_pdb = glob.glob(os.path.join(stem, '{0}_refinement_output.pdb'.format(pdb_code)))[0]
-    input_mtz = glob.glob(os.path.join(stem, '{0}_refinement_output.mtz'.format(pdb_code)))[0]
+    stem = os.path.join(run_dir, 'output_files', pdb_code)
+    input_pdb = os.path.join(stem, '{0}_refinement_output.pdb'.format(pdb_code))
+    input_mtz = os.path.join(stem, '{0}_refinement_output.mtz'.format(pdb_code))
     shutil.copyfile(input_pdb, output_pdb)
     shutil.copyfile(input_mtz, output_mtz)